added admin interface and service,, and updated dbcontext

+++ HospitalManagement.Core/Interfaces/IPatientRepository.cs
@@ -4,7 +4,7 @@ namespace HospitalManagement.Core.Interfaces;
 public interface IPatientRepository
 {
     Task<IEnumerable<Patient>> GetAllAsync();
-    Task<Patient?> GetByIdAsync(int id);
+    Task<Patient?> GetByIdAsync(long id);
     Task<Patient> CreateAsync(Patient patient);
     Task<Patient?> UpdateAsync(Patient patient);
 }
